+++ src/schemathesis/core/warnings.py
@@ -10,6 +10,7 @@ class SchemathesisWarning(str, enum.Enum):
     MISSING_DESERIALIZER = "missing_deserializer"
     UNUSED_OPENAPI_AUTH = "unused_openapi_auth"
     UNSUPPORTED_REGEX = "unsupported_regex"
+    METHOD_NOT_ALLOWED = "method_not_allowed"
 
     @classmethod
     def from_str(cls, value: str) -> SchemathesisWarning:
@@ -20,4 +21,5 @@ class SchemathesisWarning(str, enum.Enum):
             "missing_deserializer": cls.MISSING_DESERIALIZER,
             "unused_openapi_auth": cls.UNUSED_OPENAPI_AUTH,
             "unsupported_regex": cls.UNSUPPORTED_REGEX,
+            "method_not_allowed": cls.METHOD_NOT_ALLOWED,
         }[value.lower()]